requirements.txt
functions-framework
flask
google-cloud-aiplatform
protobuf